Moved functions into koda.go
