+++ client.go
@@ -1,15 +1,11 @@
 package koda
 
 import (
-	"net/url"
 	"os"
 	"os/signal"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
-
-	"gopkg.in/redis.v3"
 )
 
 var DefaultClient = NewClient(nil)
@@ -24,50 +20,6 @@ type Options struct {
 	URL         string
 	Prefix      string
 	ConnFactory func() Conn
-}
-
-func Configure(opts *Options) {
-	DefaultClient = NewClient(opts)
-}
-
-func NewClient(opts *Options) *Client {
-	if opts == nil {
-		opts = &Options{}
-	}
-
-	if opts.URL == "" {
-		opts.URL = "redis://localhost:6379"
-	}
-
-	if opts.Prefix == "" {
-		opts.Prefix = "koda"
-	}
-
-	if opts.ConnFactory == nil {
-		url, err := url.Parse(opts.URL)
-		db, _ := strconv.Atoi(url.Path)
-
-		// TODO: return err
-		if err != nil {
-			panic(err)
-		}
-
-		opts.ConnFactory = func() Conn {
-			r := redis.NewClient(&redis.Options{
-				Addr: url.Host,
-				DB:   int64(db),
-			})
-			return &GoRedisAdapter{R: r}
-		}
-	}
-
-	return &Client{
-		opts:        opts,
-		dispatchers: make(map[string]*dispatcher),
-		connPool: sync.Pool{New: func() interface{} {
-			return opts.ConnFactory()
-		}},
-	}
 }
 
 func (c *Client) newQueue(name string) *Queue {
